refactor(blank): remove unnecessary input components
