+++ packages/editor/src/plugins/fill-in-the-blanks-exercise/blank-renderer.tsx
@@ -1,7 +1,16 @@
-import { ChangeEventHandler, useContext } from 'react'
+import { cn } from '@serlo/frontend/src/helper/cn'
+import {
+  ChangeEventHandler,
+  KeyboardEvent,
+  createRef,
+  useContext,
+  useEffect,
+} from 'react'
+import { Range, Transforms } from 'slate'
+import { ReactEditor, useSelected, useSlate, useFocused } from 'slate-react'
 
-import { BlankRendererInput } from './components/blank-renderer-input'
 import { FillInTheBlanksContext } from './context/blank-context'
+import { setTextUserTypedIntoBlank } from './util/handlers'
 
 interface BlankRendererProps {
   blankId: string
@@ -11,14 +20,103 @@ interface BlankRendererProps {
 export function BlankRenderer(props: BlankRendererProps) {
   const { blankId, onChange } = props
 
+  const editor = useSlate()
+  const selected = useSelected()
+  const focused = useFocused()
+
+  // Autofocus when adding and removing a blank
+  const inputRef = createRef<HTMLInputElement>()
+  useEffect(() => {
+    // Focus input when the blank is added
+    const input = inputRef.current
+    if (input) input.focus()
+
+    // Focus editor when the blank is removed
+    return () => {
+      ReactEditor.focus(editor)
+    }
+
+    // Only run on mount
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
+  // Focus input when the blank is selected using arrow keys
+  useEffect(() => {
+    function handleDocumentKeydown() {
+      const input = inputRef.current
+      const shouldFocusInput =
+        input &&
+        document.activeElement !== input &&
+        focused &&
+        selected &&
+        editor.selection &&
+        Range.isCollapsed(editor.selection)
+      if (shouldFocusInput) input.focus()
+    }
+
+    document.addEventListener('keydown', handleDocumentKeydown)
+
+    return () => document.removeEventListener('keydown', handleDocumentKeydown)
+  }, [editor, focused, inputRef, selected])
+
   const context = useContext(FillInTheBlanksContext)
   if (context === null) return null
 
+  const feedback = context.feedbackForBlanks
+  const isAnswerCorrect = feedback.get(blankId)?.isCorrect
+  const text = context.textInBlanks.get(blankId)?.text ?? ''
+
   return (
-    <BlankRendererInput
-      blankId={blankId}
-      context={context}
-      onChange={onChange}
+    <input
+      ref={inputRef}
+      className={cn(
+        'h-[25px] resize-none rounded-full border border-brand bg-brand-50 pl-2 pr-1',
+        isAnswerCorrect && 'border-green-500',
+        isAnswerCorrect === false && 'border-red-500'
+      )}
+      size={(text.length ?? 4) + 1}
+      spellCheck={false}
+      autoCorrect="off"
+      placeholder=""
+      type="text"
+      value={text}
+      onChange={(event) => {
+        setTextUserTypedIntoBlank(context, blankId, event.target.value)
+        onChange?.(event)
+      }}
+      onKeyDown={handleMoveOut}
     />
   )
+
+  function handleMoveOut(event: KeyboardEvent<HTMLInputElement>) {
+    if (!inputRef || !inputRef.current) return
+
+    const { selectionStart, selectionEnd, value } = inputRef.current
+
+    const selectionCollapsed = selectionStart === selectionEnd
+    const caretAtRightEnd = selectionEnd === value.length
+    const caretAtLeftEnd = selectionStart === 0
+
+    // Move the selection right of the blank on arrow right
+    if (
+      event.key === 'ArrowRight' &&
+      !event.shiftKey &&
+      selectionCollapsed &&
+      caretAtRightEnd
+    ) {
+      Transforms.move(editor, { unit: 'character' })
+      ReactEditor.focus(editor)
+    }
+
+    // Move the selection left of the blank on arrow left
+    if (
+      event.key === 'ArrowLeft' &&
+      !event.shiftKey &&
+      selectionCollapsed &&
+      caretAtLeftEnd
+    ) {
+      Transforms.move(editor, { unit: 'character', reverse: true })
+      ReactEditor.focus(editor)
+    }
+  }
 }
